refactor(GiniInternalPaymentSDK): Split PaymentReviewTests into one file per suite
- Replace 829-line monolithic PaymentReviewTests.swift with 13 focused
  test files and a shared PaymentReviewTestHelpers.swift
- Promote private factory extensions to internal so they can be shared
  across all test files in the same target
- Remove duplicate factory extensions from AccessibilityTests.swift
  (BottomSheetConfiguration, ButtonConfiguration, PoweredByGiniConfiguration,
  PoweredByGiniStrings) that are now provided by the shared helpers file
- Add missing GiniUtilites import to files that reference Price

HEAL-333
